requirements.txt
geocoder==1.38.1
google_api_python_client==2.161.0
google_auth_oauthlib==1.2.1
gradio==5.19.0
ipython==8.32.0
langchain==0.3.19
langchain_core==0.3.40
langchain_google_genai==2.0.11
langgraph==0.3.1
protobuf==5.29.3
python-dotenv==1.0.1
Requests==2.32.3
typing_extensions==4.12.2
